Refactor auth and registration to handle cookies in response
Updated AuthController and RegisterUserController to delegate setting of refresh token cookies to service layer methods, passing HttpServletResponse as a parameter. Refactored UserDTO.Response to remove the cookie field, and updated AuthUserService and CreateUserService to set cookies directly on the response. Adjusted CreateRefreshToken to set cookie path and SameSite attributes based on environment. Improves separation of concerns and prepares for more flexible cookie handling.